refactor(auth): extract :client-auth + :client-auth-smartycraft from the launcher
The first real seam carved out of the launcher god-module. Auth is now two new modules: :client-auth holds the provider-agnostic AuthProvider SPI plus AbstractCachingAuthProvider (the session cache, the retry + error-translation funnel, the transient/SSL classifiers), and :client-auth-smartycraft holds SmartyCraftAuthProvider -- the SmartyCraft-specific token (AUTH_SALT + AES), V1 login request, status mapping, and TWOAUTH/2FA flow lifted out of the old core AuthService.

AuthProvider replaces IAuthService and carries AuthCapabilities(supports2FA). SmartyCraft sets supports2FA = false (its 2FA login works on the wire but breaks the game-side session), and LoginPanel now branches on the capability: a future capable provider opens the ConfirmCodeDialog path, SmartyCraft keeps the "2FA unsupported" banner. core AuthService.kt and IAuthService.kt are deleted; the provider test moves to :client-auth-smartycraft.

What stays in core: IServerProtocol + protocol models + repositories (the whole SmartyCraft wire protocol, consumed by Server/Skin/Player repositories too), SessionData, AuthStatus, AuthException/TwoFactorRequiredException. SmartycraftV1Protocol, CredentialsManager and keyring stay in the launcher; DI is rewired in place (secure + insecure variants preserved). New modules carry no Koin and depend on neither launcher nor ui -- the seam is acyclic.
