+++ client-core/src/main/kotlin/hivens/core/api/AuthException.kt
@@ -11,11 +11,10 @@ open class AuthException(
 
 /**
  * Thrown when the server returns `TWOAUTH` (account has TOTP 2FA). UI
- * catches this, prompts for the 6-digit code, follows up via
- * [hivens.core.api.interfaces.IAuthService.completeTwoFactor]. [uid] is
- * the session identifier from the TWOAUTH response (required to sign the
- * follow-up); a null [uid] means the server response was truncated and
- * the user must restart login.
+ * catches this, prompts for the 6-digit code, follows up via the auth
+ * provider's `completeTwoFactor`. [uid] is the session identifier from the
+ * TWOAUTH response (required to sign the follow-up); a null [uid] means the
+ * server response was truncated and the user must restart login.
  */
 class TwoFactorRequiredException(
     val uid: String?,
