+++ client-ui/build.gradle.kts
@@ -67,6 +67,7 @@ kotlin {
 
                 implementation(project(":client-config"))
                 implementation(project(":client-core"))
+                implementation(project(":client-auth"))
                 implementation(project(":client-launcher"))
                 implementation(project(":widget-api"))
