+++ settings.gradle.kts
@@ -64,6 +64,10 @@ rootProject.name = "nexira"
 
 include(":client-config")
 include(":client-core")
+// Auth seam carved out of the launcher god-module: :client-auth holds the
+// provider-agnostic AuthProvider SPI, :client-auth-smartycraft the SmartyCraft impl.
+include(":client-auth")
+include(":client-auth-smartycraft")
 include(":client-launcher")
 include(":client-ui")
 include(":widget-model")
